+++ rust/main/chains/hyperlane-sealevel/src/utils.rs
@@ -1,7 +1,7 @@
 use std::str::FromStr;
 
-use solana_sdk::bs58;
 use solana_sdk::pubkey::Pubkey;
+use solana_sdk::{bs58, instruction::AccountMeta};
 
 use hyperlane_core::{H256, H512};
 
@@ -30,4 +30,43 @@ pub fn decode_h512(base58: &str) -> Result<H512, HyperlaneSealevelError> {
 
 pub fn decode_pubkey(address: &str) -> Result<Pubkey, HyperlaneSealevelError> {
     Pubkey::from_str(address).map_err(Into::<HyperlaneSealevelError>::into)
+}
+
+/// Force all provided account metas to be non-signers
+pub fn force_non_signers(mut account_metas: Vec<AccountMeta>) -> Vec<AccountMeta> {
+    account_metas.iter_mut().for_each(|meta| {
+        if meta.is_signer {
+            tracing::warn!(meta = ?meta, "Forcing account meta to be non-signer");
+            meta.is_signer = false
+        }
+    });
+
+    account_metas
+}
+
+#[cfg(test)]
+mod test {
+    use crate::utils::force_non_signers;
+
+    #[test]
+    fn test_force_non_signers() {
+        use solana_sdk::instruction::AccountMeta;
+
+        let account_metas = vec![
+            AccountMeta::new_readonly([0u8; 32].into(), false),
+            AccountMeta::new_readonly([1u8; 32].into(), true),
+            AccountMeta::new([2u8; 32].into(), true),
+        ];
+
+        let account_metas = force_non_signers(account_metas.clone());
+
+        assert_eq!(
+            account_metas,
+            vec![
+                AccountMeta::new_readonly([0u8; 32].into(), false),
+                AccountMeta::new_readonly([1u8; 32].into(), false),
+                AccountMeta::new([2u8; 32].into(), false),
+            ]
+        )
+    }
 }