+++ rust/main/hyperlane-base/src/settings/base.rs
@@ -2,12 +2,14 @@ use std::{collections::HashMap, fmt::Debug, hash::Hash, sync::Arc};
 
 use eyre::{eyre, Context, Result};
 use futures_util::future::join_all;
+use tracing::warn;
+
 use hyperlane_core::{
     HyperlaneDomain, HyperlaneDomainProtocol, HyperlaneLogStore, HyperlaneProvider,
     HyperlaneSequenceAwareIndexerStoreReader, HyperlaneWatermarkedLogStore, InterchainGasPaymaster,
     Mailbox, MerkleTreeHook, MultisigIsm, SequenceAwareIndexer, ValidatorAnnounce, H256,
 };
-use tracing::warn;
+use hyperlane_operation_verifier::ApplicationOperationVerifier;
 
 use crate::{
     cursors::{CursorType, Indexable},
@@ -128,7 +130,7 @@ impl Settings {
 }
 
 /// Generate a call to ChainSetup for the given builder
-macro_rules! build_contract_fns {
+macro_rules! build_chain_conf_fns {
     ($singular:ident, $plural:ident -> $ret:ty) => {
         /// Delegates building to ChainSetup
         pub async fn $singular(
@@ -158,11 +160,12 @@ macro_rules! build_contract_fns {
 type SequenceIndexer<T> = Arc<dyn SequenceAwareIndexer<T>>;
 
 impl Settings {
-    build_contract_fns!(build_interchain_gas_paymaster, build_interchain_gas_paymasters -> dyn InterchainGasPaymaster);
-    build_contract_fns!(build_mailbox, build_mailboxes -> dyn Mailbox);
-    build_contract_fns!(build_merkle_tree_hook, build_merkle_tree_hooks -> dyn MerkleTreeHook);
-    build_contract_fns!(build_validator_announce, build_validator_announces -> dyn ValidatorAnnounce);
-    build_contract_fns!(build_provider, build_providers -> dyn HyperlaneProvider);
+    build_chain_conf_fns!(build_application_operation_verifier, build_application_operation_verifiers -> dyn ApplicationOperationVerifier);
+    build_chain_conf_fns!(build_interchain_gas_paymaster, build_interchain_gas_paymasters -> dyn InterchainGasPaymaster);
+    build_chain_conf_fns!(build_mailbox, build_mailboxes -> dyn Mailbox);
+    build_chain_conf_fns!(build_merkle_tree_hook, build_merkle_tree_hooks -> dyn MerkleTreeHook);
+    build_chain_conf_fns!(build_provider, build_providers -> dyn HyperlaneProvider);
+    build_chain_conf_fns!(build_validator_announce, build_validator_announces -> dyn ValidatorAnnounce);
 
     /// Build a contract sync for type `T` using log store `S`
     pub async fn sequenced_contract_sync<T, S>(