+++ rust/main/utils/run-locally/src/cosmos/mod.rs
@@ -31,7 +31,9 @@ use crate::logging::log;
 use crate::metrics::agent_balance_sum;
 use crate::program::Program;
 use crate::types::{AgentConfig, AgentConfigOut, HyperlaneStack};
-use crate::utils::{as_task, concat_path, stop_child, AgentHandles, TaskHandle};
+use crate::utils::{
+    as_task, concat_path, get_workspace_path, stop_child, AgentHandles, TaskHandle,
+};
 use crate::AGENT_BIN_PATH;
 pub use cli::OsmosisCLI;
 use cli::OsmosisEndpoint;
@@ -330,10 +332,12 @@ fn run_locally() {
     const TIMEOUT_SECS: u64 = 60 * 10;
     let debug = false;
 
+    let workspace_path = get_workspace_path();
+
     log!("Building rust...");
     Program::new("cargo")
         .cmd("build")
-        .working_dir("../../")
+        .working_dir(&workspace_path)
         .arg("features", "test-utils")
         .arg("bin", "relayer")
         .arg("bin", "validator")
@@ -514,7 +518,8 @@ fn run_locally() {
     // give things a chance to fully start.
     sleep(Duration::from_secs(10));
 
-    let starting_relayer_balance: f64 = agent_balance_sum(hpl_rly_metrics_port).unwrap();
+    let starting_relayer_balance: f64 =
+        agent_balance_sum(hpl_rly_metrics_port).expect("Failed to get relayer agent balance");
 
     // dispatch the second batch of messages (after agents start)
     dispatched_messages += dispatch(&osmosisd, linker, &nodes);