+++ rust/main/utils/run-locally/src/utils.rs
@@ -2,13 +2,10 @@ use std::collections::HashMap;
 use std::fs::File;
 use std::io::{self, BufRead};
 use std::path::{Path, PathBuf};
-use std::process::Child;
+use std::process::{Child, Command};
 use std::sync::{Arc, Mutex};
 use std::thread::JoinHandle;
 
-use fuels::types::bech32::Bech32ContractId;
-use fuels::types::ContractId;
-use hyperlane_cosmwasm_interface::types::bech32_decode;
 use nix::libc::pid_t;
 use nix::sys::signal;
 use nix::sys::signal::Signal;
@@ -122,26 +119,88 @@ pub fn stop_child(child: &mut Child) {
     };
 }
 
-pub fn get_matching_lines(file: &File, search_strings: &[&str]) -> HashMap<String, u32> {
+/// Given a Vec<Vec<&str>>,
+/// for each Vec<&str>, count how many lines in the file
+/// matches all the &str in that Vec.
+/// Store this count in a hashmap where the key is the vector
+/// Vec<&str>
+/// and return this hashmap.
+pub fn get_matching_lines<'a>(
+    file: &File,
+    search_strings: Vec<Vec<&'a str>>,
+) -> HashMap<Vec<&'a str>, u32> {
     let reader = io::BufReader::new(file);
     let mut matches = HashMap::new();
 
     let mut lines = reader.lines();
     while let Some(Ok(line)) = lines.next() {
-        search_strings.iter().for_each(|search_string| {
-            if line.contains(search_string) {
-                let count = matches.entry(search_string.to_string()).or_insert(0);
+        search_strings.iter().for_each(|search_string_vec| {
+            if search_string_vec
+                .iter()
+                .map(|search_string| line.contains(search_string))
+                .all(|x| x)
+            {
+                let count = matches.entry(search_string_vec.clone()).or_insert(0);
                 *count += 1;
             }
         });
     }
     matches
 }
-
-pub fn fuel_to_hex_addr(addr: &Bech32ContractId) -> String {
-    format!("0x{}", ContractId::from(addr).to_string())
+#[cfg(feature = "fuel")]
+pub fn fuel_to_hex_addr(addr: &fuels::types::bech32::Bech32ContractId) -> String {
+    format!("0x{}", fuels::types::ContractId::from(addr).to_string())
 }
 
+#[cfg(feature = "cosmos")]
 pub fn cw_to_hex_addr(addr: &str) -> String {
-    format!("0x{}", hex::encode(bech32_decode(addr).unwrap()))
+    format!(
+        "0x{}",
+        hex::encode(hyperlane_cosmwasm_interface::types::bech32_decode(addr).unwrap())
+    )
+}
+
+/// Returns absolute path to rust workspace
+/// `/<...>/hyperlane-monorepo/rust/main`.
+/// This allows us to have a more reliable way of generating
+/// relative paths such path to sealevel directory
+pub fn get_workspace_path() -> PathBuf {
+    let output = Command::new(env!("CARGO"))
+        .arg("locate-project")
+        .arg("--workspace")
+        .arg("--message-format=plain")
+        .output()
+        .expect("Failed to get workspace path")
+        .stdout;
+    let path_str = String::from_utf8(output).expect("Failed to parse workspace path");
+    let mut workspace_path = PathBuf::from(path_str);
+    // pop Cargo.toml from path
+    workspace_path.pop();
+    workspace_path
+}
+
+/// Returns absolute path to sealevel directory
+/// `/<...>/hyperlane-monorepo/rust/sealevel`
+#[cfg(feature = "sealevel")]
+pub fn get_sealevel_path(workspace_path: &Path) -> PathBuf {
+    concat_path(
+        workspace_path
+            .parent()
+            .expect("workspace path has no parent"),
+        "sealevel",
+    )
+}
+
+/// Returns absolute path to typescript infra directory
+/// `/<...>/hyperlane-monorepo/typescript/infra`
+pub fn get_ts_infra_path() -> PathBuf {
+    let output = Command::new("git")
+        .arg("rev-parse")
+        .arg("--show-toplevel")
+        .output()
+        .expect("Failed to get git workspace path")
+        .stdout;
+    let path_str = String::from_utf8(output).expect("Failed to parse workspace path");
+    let git_workspace_path = PathBuf::from(path_str.trim());
+    concat_path(git_workspace_path, "typescript/infra")
 }