+++ typescript/infra/config/environments/mainnet3/funding.ts
@@ -10,7 +10,7 @@ export const keyFunderConfig: KeyFunderConfig<
 > = {
   docker: {
     repo: 'gcr.io/abacus-labs-dev/hyperlane-monorepo',
-    tag: '95deca3-20250120-103554',
+    tag: '114fd57-20250218-171940',
   },
   // We're currently using the same deployer/key funder key as mainnet2.
   // To minimize nonce clobbering we offset the key funder cron
@@ -26,6 +26,7 @@ export const keyFunderConfig: KeyFunderConfig<
   },
   // desired balance config, must be set for each chain
   desiredBalancePerChain: {
+    abstract: '0',
     // acala: '5',
     ancient8: '0.5',
     alephzeroevmmainnet: '100',
@@ -33,15 +34,18 @@ export const keyFunderConfig: KeyFunderConfig<
     appchain: '0.05',
     arbitrum: '0.5',
     arbitrumnova: '0.05',
+    arcadia: '0.05',
     artela: '2',
     arthera: '0.1',
     astar: '100',
     astarzkevm: '0.05',
     aurora: '0.05',
+    bouncebit: '1',
     flame: '1',
     avalanche: '5',
     b3: '0.05',
     base: '0.5',
+    berachain: '0.05',
     bitlayer: '0.002',
     blast: '0.2',
     bob: '0.2',
@@ -70,11 +74,13 @@ export const keyFunderConfig: KeyFunderConfig<
     // fractal: '500',
     fraxtal: '0.2',
     fusemainnet: '20',
+    glue: '50',
     gnosis: '5',
     gravity: '500',
     guru: '100',
     harmony: '500',
     hemi: '0.05',
+    hyperevm: '1',
     immutablezkevmmainnet: '25',
     inevm: '3',
     ink: '0.05',
@@ -87,6 +93,7 @@ export const keyFunderConfig: KeyFunderConfig<
     lumiaprism: '1',
     mantapacific: '0.2',
     mantle: '20',
+    matchain: '0.05',
     merlin: '0.002',
     metal: '0.05',
     metis: '3',
@@ -108,6 +115,7 @@ export const keyFunderConfig: KeyFunderConfig<
     real: '0.1',
     redstone: '0.2',
     rivalz: '0.05',
+    ronin: '10',
     rootstockmainnet: '0.002',
     sanko: '2',
     scroll: '0.5',
@@ -116,9 +124,11 @@ export const keyFunderConfig: KeyFunderConfig<
     snaxchain: '0.05',
     soneium: '0.05',
     sonic: '0.5',
+    sophon: '10',
+    story: '10',
     // ignore non-evm chains
     stride: '0',
-    // subtensor: '0.05',
+    subtensor: '0.05',
     superseed: '0.05',
     superpositionmainnet: '0.05',
     swell: '0.05',
@@ -129,6 +139,7 @@ export const keyFunderConfig: KeyFunderConfig<
     treasure: '900',
     trumpchain: '0.25',
     unichain: '0.05',
+    unitzero: '50',
     // temporarily low until we're able to fund more
     vana: '0.001',
     viction: '3',
@@ -149,6 +160,7 @@ export const keyFunderConfig: KeyFunderConfig<
     solanamainnet: '0',
     eclipsemainnet: '0',
     soon: '0',
+    sonicsvm: '0',
   },
   // if not set, keyfunder defaults to 0
   desiredKathyBalancePerChain: {
@@ -205,6 +217,7 @@ export const keyFunderConfig: KeyFunderConfig<
     eclipsemainnet: '0',
     solanamainnet: '0',
     soon: '0',
+    sonicsvm: '0',
   },
   // if not set, keyfunder defaults to using desired balance * 0.2 as the threshold
   igpClaimThresholdPerChain: {
@@ -261,5 +274,6 @@ export const keyFunderConfig: KeyFunderConfig<
     eclipsemainnet: '0',
     solanamainnet: '0',
     soon: '0',
+    sonicsvm: '0',
   },
 };